Initial fixes to get to 10.10 and beyond.

+++ Classes/LINBasicPerformer.m
@@ -42,7 +42,7 @@ static id sharedInstance = nil;
 	
 	do {
 		sequenceNumber++;
-		temporaryPath = [NSString stringWithFormat:@"%d-%d-%d.%@", [[NSProcessInfo processInfo] processIdentifier], (NSInteger)[NSDate timeIntervalSinceReferenceDate], sequenceNumber, @"Lingon"];
+		temporaryPath = [NSString stringWithFormat:@"%d-%ld-%ld.%@", [[NSProcessInfo processInfo] processIdentifier], (long)[NSDate timeIntervalSinceReferenceDate], (long)sequenceNumber, @"Lingon"];
 		temporaryPath = [NSTemporaryDirectory() stringByAppendingPathComponent:temporaryPath];
 	} 
 	while ([[NSFileManager defaultManager] fileExistsAtPath:temporaryPath]);
